Split hooks to hooks-in-components/
