+++ tests/hooks-in-components/watching-nested-stores.spec.tsx
@@ -1,0 +1,357 @@
+import React from "react"
+import { act, render, screen, fireEvent } from "@testing-library/react"
+
+import { InnerStore, useInnerState, useInnerWatch } from "../../src"
+
+import { CounterComponent, expectCounts, withinNth } from "./common"
+
+describe("watching nested stores", () => {
+  abstract class AppState {
+    public abstract counts: ReadonlyArray<InnerStore<number>>
+
+    public static sum({ counts }: AppState): number {
+      return counts.reduce((acc, count) => acc + count.getState(), 0)
+    }
+  }
+
+  interface AppProps {
+    store: InnerStore<AppState>
+    onRender: VoidFunction
+    onCounterRender: React.Dispatch<number>
+  }
+
+  const GenericApp: React.VFC<
+    {
+      moreThanTen: boolean
+      lessThanTwenty: boolean
+    } & AppProps
+  > = ({ moreThanTen, lessThanTwenty, store, onRender, onCounterRender }) => {
+    const [state, setState] = useInnerState(store)
+
+    onRender()
+
+    return (
+      <>
+        {moreThanTen && <span>more than ten</span>}
+        {lessThanTwenty && <span>less than twenty</span>}
+
+        <button
+          type="button"
+          data-testid="add-counter"
+          onClick={() => {
+            setState({
+              ...state,
+              counts: [...state.counts, InnerStore.of(0)],
+            })
+          }}
+        />
+        <button
+          type="button"
+          data-testid="reset-counters"
+          onClick={() => {
+            state.counts.forEach((count) => {
+              count.setState(0)
+
+              return count
+            })
+          }}
+        />
+        <button
+          type="button"
+          data-testid="increment-all"
+          onClick={() => {
+            store.setState((current) => {
+              current.counts.forEach((count) => {
+                count.setState((x) => x + 1)
+
+                return count
+              })
+
+              return current
+            })
+          }}
+        />
+        {state.counts.map((count, index) => (
+          <CounterComponent
+            key={count.key}
+            count={count}
+            onRender={() => onCounterRender(index)}
+          />
+        ))}
+      </>
+    )
+  }
+
+  const SingleWatcherApp: React.VFC<AppProps> = (props) => {
+    const [moreThanTen, lessThanTwenty] = useInnerWatch(
+      () => {
+        const count = AppState.sum(props.store.getState())
+
+        return [count > 10, count < 20]
+      },
+      ([left1, right1], [left2, right2]) => {
+        return left1 === left2 && right1 === right2
+      },
+    )
+
+    return (
+      <GenericApp
+        moreThanTen={moreThanTen}
+        lessThanTwenty={lessThanTwenty}
+        {...props}
+      />
+    )
+  }
+
+  const SingleMemoizedWatcherApp: React.VFC<AppProps> = (props) => {
+    const [moreThanTen, lessThanTwenty] = useInnerWatch(
+      React.useCallback(() => {
+        const count = AppState.sum(props.store.getState())
+
+        return [count > 10, count < 20]
+      }, [props.store]),
+      React.useCallback(([left1, right1], [left2, right2]) => {
+        return left1 === left2 && right1 === right2
+      }, []),
+    )
+
+    return (
+      <GenericApp
+        moreThanTen={moreThanTen}
+        lessThanTwenty={lessThanTwenty}
+        {...props}
+      />
+    )
+  }
+
+  const MultipleWatchersApp: React.VFC<AppProps> = (props) => {
+    const moreThanTen = useInnerWatch(() => {
+      const count = props.store.getState(AppState.sum)
+
+      return count > 10
+    })
+    const lessThanTwenty = useInnerWatch(() => {
+      const count = AppState.sum(props.store.getState())
+
+      return count < 20
+    })
+
+    return (
+      <GenericApp
+        moreThanTen={moreThanTen}
+        lessThanTwenty={lessThanTwenty}
+        {...props}
+      />
+    )
+  }
+
+  const MultipleMemoizedWatchersApp: React.VFC<AppProps> = (props) => {
+    const moreThanTen = useInnerWatch(
+      React.useCallback(() => {
+        const count = props.store.getState(AppState.sum)
+
+        return count > 10
+      }, [props.store]),
+    )
+    const lessThanTwenty = useInnerWatch(
+      React.useCallback(() => {
+        const count = AppState.sum(props.store.getState())
+
+        return count < 20
+      }, [props.store]),
+    )
+
+    return (
+      <GenericApp
+        moreThanTen={moreThanTen}
+        lessThanTwenty={lessThanTwenty}
+        {...props}
+      />
+    )
+  }
+
+  it.each([
+    ["single watcher", SingleWatcherApp],
+    ["single memoized watcher", SingleMemoizedWatcherApp],
+    ["multiple watchers", MultipleWatchersApp],
+    ["multiple memoized watchers", MultipleMemoizedWatchersApp],
+  ])("watches nested stores with %s", (_, App) => {
+    const store = InnerStore.of<AppState>({
+      counts: [],
+    })
+    const onRender = jest.fn()
+    const onCounterRender = jest.fn()
+
+    render(
+      <App
+        store={store}
+        onRender={onRender}
+        onCounterRender={onCounterRender}
+      />,
+    )
+
+    // initial render and watcher setup
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(0)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([])
+    jest.clearAllMocks()
+
+    // add first counter
+    fireEvent.click(screen.getByTestId("add-counter"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([0])
+    jest.clearAllMocks()
+
+    // increment first counter
+    fireEvent.click(withinNth("counter", 0).getByTestId("increment"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([1])
+    jest.clearAllMocks()
+
+    // add second counter
+    fireEvent.click(screen.getByTestId("add-counter"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 1)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([1, 0])
+    jest.clearAllMocks()
+
+    // increment all counters
+    fireEvent.click(screen.getByTestId("increment-all"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).toHaveBeenCalledTimes(2)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(onCounterRender).toHaveBeenNthCalledWith(2, 1)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([2, 1])
+    jest.clearAllMocks()
+
+    // add third counter
+    fireEvent.click(screen.getByTestId("add-counter"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 2)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([2, 1, 0])
+    jest.clearAllMocks()
+
+    // reset counters
+    fireEvent.click(screen.getByTestId("reset-counters"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).toHaveBeenCalledTimes(2)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(onCounterRender).toHaveBeenNthCalledWith(2, 1)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([0, 0, 0])
+    jest.clearAllMocks()
+
+    // add fourth counter from the outside
+    act(() => {
+      store.setState((state) => ({
+        ...state,
+        counts: [...state.counts, InnerStore.of(9)],
+      }))
+    })
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 3)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([0, 0, 0, 9])
+    jest.clearAllMocks()
+
+    // increment all counters
+    fireEvent.click(screen.getByTestId("increment-all"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(4)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(onCounterRender).toHaveBeenNthCalledWith(2, 1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(3, 2)
+    expect(onCounterRender).toHaveBeenNthCalledWith(4, 3)
+    expect(screen.queryByText("more than ten")).toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([1, 1, 1, 10])
+    jest.clearAllMocks()
+
+    // add fifth counter
+    fireEvent.click(screen.getByTestId("add-counter"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 4)
+    expect(screen.queryByText("more than ten")).toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([1, 1, 1, 10, 0])
+    jest.clearAllMocks()
+
+    // increment all counters
+    fireEvent.click(screen.getByTestId("increment-all"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).toHaveBeenCalledTimes(5)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(onCounterRender).toHaveBeenNthCalledWith(2, 1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(3, 2)
+    expect(onCounterRender).toHaveBeenNthCalledWith(4, 3)
+    expect(onCounterRender).toHaveBeenNthCalledWith(5, 4)
+    expect(screen.queryByText("more than ten")).toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([2, 2, 2, 11, 1])
+    jest.clearAllMocks()
+
+    // increment fifth counter
+    fireEvent.click(withinNth("counter", 4).getByTestId("increment"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 4)
+    expect(screen.queryByText("more than ten")).toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([2, 2, 2, 11, 2])
+    jest.clearAllMocks()
+
+    // increment fourth counter
+    fireEvent.click(withinNth("counter", 3).getByTestId("increment"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 3)
+    expect(screen.queryByText("more than ten")).toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).not.toBeInTheDocument()
+    expectCounts([2, 2, 2, 12, 2])
+    jest.clearAllMocks()
+
+    // reset all counters
+    fireEvent.click(screen.getByTestId("reset-counters"))
+    expect(onRender).toHaveBeenCalledTimes(1)
+    expect(onCounterRender).toHaveBeenCalledTimes(5)
+    expect(onCounterRender).toHaveBeenNthCalledWith(1, 0)
+    expect(onCounterRender).toHaveBeenNthCalledWith(2, 1)
+    expect(onCounterRender).toHaveBeenNthCalledWith(3, 2)
+    expect(onCounterRender).toHaveBeenNthCalledWith(4, 3)
+    expect(onCounterRender).toHaveBeenNthCalledWith(5, 4)
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([0, 0, 0, 0, 0])
+    jest.clearAllMocks()
+
+    // reset all counters again
+    fireEvent.click(screen.getByTestId("reset-counters"))
+    expect(onRender).not.toHaveBeenCalled()
+    expect(onCounterRender).not.toHaveBeenCalled()
+    expect(screen.queryByText("more than ten")).not.toBeInTheDocument()
+    expect(screen.queryByText("less than twenty")).toBeInTheDocument()
+    expectCounts([0, 0, 0, 0, 0])
+  })
+})
